linked create event endpoint
